Support source subfolders in the makefile

+++ Makefile
@@ -25,31 +25,35 @@ CPPFLAGS := -DMIPS -DF3DEX_GBI_2 -DF3DEX_GBI_PL -DGBI_DOWHILE -I include -I incl
             -I mm-decomp/include -I mm-decomp/src -I mm-decomp/extracted/n64-us -idirafter include/libc -idirafter mm-decomp/include/libc
 LDFLAGS  := -nostdlib -T $(LDSCRIPT) -Map $(BUILD_DIR)/mod.map --unresolved-symbols=ignore-all --emit-relocs -e 0 --no-nmagic -gc-sections
 
-C_SRCS := $(wildcard src/*.c)
+rwildcard = $(foreach d,$(wildcard $(1:=/*)),$(call rwildcard,$d,$2) $(filter $(subst *,%,$2),$d))
+getdirs = $(sort $(dir $(1)))
+
+C_SRCS := $(call rwildcard,src,*.c)
 C_OBJS := $(addprefix $(BUILD_DIR)/, $(C_SRCS:.c=.o))
 C_DEPS := $(addprefix $(BUILD_DIR)/, $(C_SRCS:.c=.d))
 
 ALL_OBJS := $(C_OBJS)
 ALL_DEPS := $(C_DEPS)
+BUILD_DIRS := $(call getdirs,$(ALL_OBJS))
 
 all: $(TARGET)
 
 $(TARGET): $(ALL_OBJS) $(LDSCRIPT) | $(BUILD_DIR)
 	$(LD) $(ALL_OBJS) $(LDFLAGS) -o $@
 
-$(BUILD_DIR) $(BUILD_DIR)/src:
+$(BUILD_DIR) $(BUILD_DIRS):
 ifeq ($(OS),Windows_NT)
-	mkdir $(subst /,\,$@)
+	if not exist "$(subst /,\,$@)" mkdir "$(subst /,\,$@)"
 else
 	mkdir -p $@
 endif
 
-$(C_OBJS): $(BUILD_DIR)/%.o : %.c | $(BUILD_DIR) $(BUILD_DIR)/src
+$(C_OBJS): $(BUILD_DIR)/%.o : %.c | $(BUILD_DIRS)
 	$(CC) $(CFLAGS) $(CPPFLAGS) $< -MMD -MF $(@:.o=.d) -c -o $@
 
 clean:
 ifeq ($(OS),Windows_NT)
-	rmdir /S /Q $(BUILD_DIR)
+	if exist $(BUILD_DIR) rmdir /S /Q $(BUILD_DIR)
 else
 	rm -rf $(BUILD_DIR)
 endif
@@ -57,3 +61,6 @@ endif
 -include $(ALL_DEPS)
 
 .PHONY: clean all
+
+# Print target for debugging
+print-% : ; $(info $* is a $(flavor $*) variable set to [$($*)]) @true
